Refactor selector
