+++ src/features/HomePage/index.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { selectHasUnreadMessages } from '@/features/Chat/selectors';
 import { useAppSelector } from '@/store';
 import { useGetLayoutMode } from '@/hooks/useGetLayoutMode';
-import { selectMentorProfile } from '../MentorPage/mentorPageApi';
+import { selectMyMentorProfile } from '../MentorPage/selectors';
 
 import Announcements from './components/Announcements';
 import Background from '@/static/img/mountain-background.svg';
@@ -12,10 +12,9 @@ import FindMentor from './components/FindMentor';
 import Info from './components/Info';
 import NewestMentors from './components/NewestMentors';
 import NewMessages from './components/NewMessages';
+import ProfileWidget from './components/ProfileWidget';
 
 import { OUTER_HORIZONTAL_MARGIN } from '@/components/constants';
-
-import ProfileWidget from './components/ProfileWidget';
 
 import PageWithTransition from '@/components/PageWithTransition';
 import Welcome from './components/Welcome';
@@ -23,7 +22,7 @@ import Welcome from './components/Welcome';
 const HomePage = () => {
   const hasUnreadMessages = useAppSelector(selectHasUnreadMessages);
   const { isTablet } = useGetLayoutMode();
-  const mentor = useAppSelector(selectMentorProfile);
+  const mentor = useAppSelector(selectMyMentorProfile);
 
   return isTablet ? (
     <PageWithTransition>
@@ -42,7 +41,7 @@ const HomePage = () => {
       <MiddleContainer>
         <LeftMiddleContainer>
           {hasUnreadMessages ? <NewMessages /> : <Welcome />}
-          {mentor && <Announcements />}
+          {!mentor && <Announcements />}
           {mentor && <ProfileWidget mentor={mentor} />}
         </LeftMiddleContainer>
         <RightMiddleContainer>
